+++ src/Data/Fields/Schema/FieldMultiasset.php
@@ -6,6 +6,8 @@ namespace Storyblok\ManagementApi\Data\Fields\Schema;
 
 class FieldMultiasset extends FieldGeneric
 {
+    use FieldNamedConstructor;
+
     public const TYPE = "multiasset";
 
     /**