+++ src/Data/Fields/Schema/FieldOptions.php
@@ -6,6 +6,8 @@ namespace Storyblok\ManagementApi\Data\Fields\Schema;
 
 class FieldOptions extends FieldGeneric
 {
+    use FieldNamedConstructor;
+
     public const TYPE = "options";
 
     /**