+++ src/Data/Fields/Schema/FieldTable.php
@@ -6,6 +6,8 @@ namespace Storyblok\ManagementApi\Data\Fields\Schema;
 
 class FieldTable extends FieldGeneric
 {
+    use FieldNamedConstructor;
+
     public const TYPE = "table";
 
     /**